split StudyAffirmation component
